Rename SoqlQuery to FluentQuery for clarity
