+++ spec/fluent-query/filters/soql-location-filter.spec.ts
@@ -1,8 +1,8 @@
 import { createMock } from 'ts-auto-mock';
-import { ISodaResource } from "../../src/client/isodaresource";
-import { Location } from "../../src/datatypes/location";
-import { SoqlQuery } from "../../src/soql-query/soql-query";
-import { IQueryable } from '../../src/soql-query/iqueryable';
+import { ISodaResource } from "../../../src/client/isodaresource";
+import { Location } from "../../../src/datatypes/location";
+import { FluentQuery } from "../../../src/fluent-query/fluent-query";
+import { IQueryable } from '../../../src/fluent-query/iqueryable';
 
 describe("SoqlLocationFilter", () => {
   interface ITestInterface {
@@ -12,7 +12,7 @@ describe("SoqlLocationFilter", () => {
   }
 
   const mockResource: ISodaResource<ITestInterface> = createMock<ISodaResource<ITestInterface>>();
-  const query: IQueryable<ITestInterface> = new SoqlQuery<ITestInterface>(mockResource);
+  const query: IQueryable<ITestInterface> = new FluentQuery<ITestInterface>(mockResource);
 
   it("should generate simple function query", () => {
     const generatedQuery = query.location(x => x.location)
