+++ src/fluent-query/filters/ilocationfilter.ts
@@ -1,6 +1,6 @@
-import { Location } from '../datatypes/location';
-import { Meters } from '../datatypes/metres';
-import { IFilteredQueryable } from './ifilteredqueryable';
+import { Location } from '../../datatypes/location';
+import { Meters } from '../../datatypes/metres';
+import { IFilteredQueryable } from '../ifilteredqueryable';
 
 export interface ILocationFilter<TEntity> {
   withinBox(start: Location, end: Location): IFilteredQueryable<TEntity>;
