+++ src/fluent-query/filters/soql-geometry-filter.ts
@@ -1,17 +1,15 @@
 import { Geometry, MultiPolygon, Point } from 'geojson';
-import { Meters } from '../datatypes/metres';
-import { Column } from '../soql-query-builder/clauses/column';
-import { Intersects } from '../soql-query-builder/clauses/where/functions/intersects';
-import { WithinBox } from '../soql-query-builder/clauses/where/functions/within-box';
-import { WithinCircle } from '../soql-query-builder/clauses/where/functions/within-circle';
-import { WithinPolygon } from '../soql-query-builder/clauses/where/functions/within-polygon';
-import { Operator } from '../soql-query-builder/clauses/where/operator';
-import { WhereOperator } from '../soql-query-builder/clauses/where/where-operator';
-import { IFilteredQueryable } from './ifilteredqueryable';
+import { Meters } from '../../datatypes/metres';
+import { Column } from '../../soql-query-builder/clauses/column';
+import { Intersects } from '../../soql-query-builder/clauses/where/functions/intersects';
+import { WithinBox } from '../../soql-query-builder/clauses/where/functions/within-box';
+import { WithinCircle } from '../../soql-query-builder/clauses/where/functions/within-circle';
+import { WithinPolygon } from '../../soql-query-builder/clauses/where/functions/within-polygon';
+import { IFilteredQueryable } from '../ifilteredqueryable';
+import { IInternalQuery } from '../iinternalquery';
 import { IGeometryFilter } from './igeometryfilter';
-import { IInternalQuery } from './iinternalquery';
 
-export class SoqlOrGeometryFilter<TEntity> implements IGeometryFilter<TEntity> {
+export class SoqlGeometryFilter<TEntity> implements IGeometryFilter<TEntity> {
 
   public constructor(protected readonly query: IInternalQuery<TEntity>, protected readonly column: Column) {
     if (!query) {
@@ -28,7 +26,7 @@ export class SoqlOrGeometryFilter<TEntity> implements IGeometryFilter<TEntity> {
     }
 
     const filter = new Intersects(this.column, geometry);
-    return this.query.addFilter(new WhereOperator(Operator.Or), filter);
+    return this.query.addFilter(filter);
   }
 
   public withinCircle(point: Point, radius: Meters): IFilteredQueryable<TEntity> {
@@ -40,7 +38,7 @@ export class SoqlOrGeometryFilter<TEntity> implements IGeometryFilter<TEntity> {
     }
 
     const filter = new WithinCircle(this.column, point, radius);
-    return this.query.addFilter(new WhereOperator(Operator.Or), filter);
+    return this.query.addFilter(filter);
   }
 
   public withinBox(start: Point, end: Point): IFilteredQueryable<TEntity> {
@@ -52,7 +50,7 @@ export class SoqlOrGeometryFilter<TEntity> implements IGeometryFilter<TEntity> {
     }
 
     const filter = new WithinBox(this.column, start, end);
-    return this.query.addFilter(new WhereOperator(Operator.Or), filter);
+    return this.query.addFilter(filter);
   }
 
   public withinPolygon(multiPolygon: MultiPolygon): IFilteredQueryable<TEntity> {
@@ -61,6 +59,6 @@ export class SoqlOrGeometryFilter<TEntity> implements IGeometryFilter<TEntity> {
     }
 
     const filter = new WithinPolygon(this.column, multiPolygon);
-    return this.query.addFilter(new WhereOperator(Operator.Or), filter);
+    return this.query.addFilter(filter);
   }
 }
