+++ src/fluent-query/filters/soql-location-filter.ts
@@ -1,15 +1,13 @@
-import { Location } from '../datatypes/location';
-import { Meters } from '../datatypes/metres';
-import { Column } from '../soql-query-builder/clauses/column';
-import { WithinBox } from '../soql-query-builder/clauses/where/functions/within-box';
-import { WithinCircle } from '../soql-query-builder/clauses/where/functions/within-circle';
-import { IFilteredQueryable } from './ifilteredqueryable';
-import { IInternalQuery } from './iinternalquery';
+import { Location } from '../../datatypes/location';
+import { Meters } from '../../datatypes/metres';
+import { Column } from '../../soql-query-builder/clauses/column';
+import { WithinBox } from '../../soql-query-builder/clauses/where/functions/within-box';
+import { WithinCircle } from '../../soql-query-builder/clauses/where/functions/within-circle';
+import { IFilteredQueryable } from '../ifilteredqueryable';
+import { IInternalQuery } from '../iinternalquery';
 import { ILocationFilter } from './ilocationfilter';
-import { WhereOperator } from '../soql-query-builder/clauses/where/where-operator';
-import { Operator } from '../soql-query-builder/clauses/where/operator';
 
-export class SoqlOrLocationFilter<TEntity> implements ILocationFilter<TEntity> {
+export class SoqlLocationFilter<TEntity> implements ILocationFilter<TEntity> {
 
   public constructor(protected readonly query: IInternalQuery<TEntity>, protected readonly column: Column) {
     if (!query) {
@@ -29,7 +27,7 @@ export class SoqlOrLocationFilter<TEntity> implements ILocationFilter<TEntity> {
     }
 
     const filter = new WithinCircle(this.column, location, radius);
-    return this.query.addFilter(new WhereOperator(Operator.Or), filter);
+    return this.query.addFilter(filter);
   }
 
   public withinBox(start: Location, end: Location): IFilteredQueryable<TEntity> {
@@ -41,6 +39,6 @@ export class SoqlOrLocationFilter<TEntity> implements ILocationFilter<TEntity> {
     }
 
     const filter = new WithinBox(this.column, start, end);
-    return this.query.addFilter(new WhereOperator(Operator.Or), filter);
+    return this.query.addFilter(filter);
   }
 }
