+++ src/fluent-query/filters/soql-or-location-filter.ts
@@ -1,13 +1,15 @@
-import { Location } from '../datatypes/location';
-import { Meters } from '../datatypes/metres';
-import { Column } from '../soql-query-builder/clauses/column';
-import { WithinBox } from '../soql-query-builder/clauses/where/functions/within-box';
-import { WithinCircle } from '../soql-query-builder/clauses/where/functions/within-circle';
-import { IFilteredQueryable } from './ifilteredqueryable';
-import { IInternalQuery } from './iinternalquery';
+import { Location } from '../../datatypes/location';
+import { Meters } from '../../datatypes/metres';
+import { Column } from '../../soql-query-builder/clauses/column';
+import { WithinBox } from '../../soql-query-builder/clauses/where/functions/within-box';
+import { WithinCircle } from '../../soql-query-builder/clauses/where/functions/within-circle';
+import { Operator } from '../../soql-query-builder/clauses/where/operator';
+import { WhereOperator } from '../../soql-query-builder/clauses/where/where-operator';
+import { IFilteredQueryable } from '../ifilteredqueryable';
+import { IInternalQuery } from '../iinternalquery';
 import { ILocationFilter } from './ilocationfilter';
 
-export class SoqlLocationFilter<TEntity> implements ILocationFilter<TEntity> {
+export class SoqlOrLocationFilter<TEntity> implements ILocationFilter<TEntity> {
 
   public constructor(protected readonly query: IInternalQuery<TEntity>, protected readonly column: Column) {
     if (!query) {
@@ -27,7 +29,7 @@ export class SoqlLocationFilter<TEntity> implements ILocationFilter<TEntity> {
     }
 
     const filter = new WithinCircle(this.column, location, radius);
-    return this.query.addFilter(filter);
+    return this.query.addFilter(new WhereOperator(Operator.Or), filter);
   }
 
   public withinBox(start: Location, end: Location): IFilteredQueryable<TEntity> {
@@ -39,6 +41,6 @@ export class SoqlLocationFilter<TEntity> implements ILocationFilter<TEntity> {
     }
 
     const filter = new WithinBox(this.column, start, end);
-    return this.query.addFilter(filter);
+    return this.query.addFilter(new WhereOperator(Operator.Or), filter);
   }
 }
