+++ src/fluent-query/iinternalquery.ts
